+++ CHANGELOG.md
@@ -27,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Line3D: skew_to: Verifies if two Line3D are skew
 * LineSegment3D.line_interserctions
 * New Class wires.Ellipse2D
-* Ellipse2D.point_belongs()
+* Ellipse2D.point_over_ellipse()
 * Ellipse2D.line_intersections()
 * Ellipse2D.linesegment_intersections()
 * Ellipse2D.discretization_points()
@@ -87,7 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Line2D: sort_points_along_line
 * Line3D: sort_points_along_line
 * ArcEllipse2D: length, point_belongs, abscissa, bounding_rectangle, straight_line_area, discretization_points, reverse
-* Ellipse2D.point_belongs()
+* Ellipse2D.point_over_ellipse()
 * Ellipse2D.line_intersections()
 * Ellipse2D.linesegment_intersections()
 * Ellipse2D.discretization_points()